requirements.txt
requests[socks]
beautifulsoup4
PyYAML